Animation: use uint16_t for the index_t
the original data is all just an ieWord and we have no use to go higher in this scenario
also fixes crashing due to bad FLTable

+++ gemrb/core/Animation.h
@@ -37,7 +37,7 @@ namespace GemRB {
 
 class GEM_EXPORT Animation {
 public:
-	using index_t = size_t;
+	using index_t = uint16_t;
 
 	bool endReached;
 	index_t frameIdx;
